Replace the Dune API calls by the Liquity stats
