+++ frontend/app/src/env.ts
@@ -115,16 +115,8 @@ export const EnvSchema = v.pipe(
       v.optional(v.string(), ""),
       v.transform((value) => value.trim() || null),
     ),
-    DUNE_API_KEY: v.optional(v.union([v.null(), v.string()]), null),
-    DUNE_POOLS_APY_URL: v.optional(
-      v.union([
-        v.null(),
-        v.pipe(v.string(), v.url()),
-      ]),
-      null,
-    ),
     KNOWN_INITIATIVES_URL: v.optional(v.pipe(v.string(), v.url())),
-    LIQUITY_STATS_URL: v.pipe(v.string(), v.url()),
+    LIQUITY_STATS_URL: v.optional(v.pipe(v.string(), v.url())),
     SAFE_API_URL: v.optional(v.pipe(v.string(), v.url())),
     SUBGRAPH_URL: v.pipe(v.string(), v.url()),
     VERCEL_ANALYTICS: v.optional(vEnvFlag(), "false"),
@@ -292,8 +284,6 @@ const parsedEnv = v.safeParse(EnvSchema, {
   DELEGATE_AUTO: process.env.NEXT_PUBLIC_DELEGATE_AUTO,
   DEMO_MODE: process.env.NEXT_PUBLIC_DEMO_MODE,
   DEPLOYMENT_FLAVOR: process.env.NEXT_PUBLIC_DEPLOYMENT_FLAVOR,
-  DUNE_API_KEY: process.env.NEXT_PUBLIC_DUNE_API_KEY,
-  DUNE_POOLS_APY_URL: process.env.NEXT_PUBLIC_DUNE_POOLS_APY_URL,
   KNOWN_INITIATIVES_URL: process.env.NEXT_PUBLIC_KNOWN_INITIATIVES_URL,
   LIQUITY_STATS_URL: process.env.NEXT_PUBLIC_LIQUITY_STATS_URL,
   SAFE_API_URL: process.env.NEXT_PUBLIC_SAFE_API_URL,
@@ -400,8 +390,6 @@ export const {
   DELEGATE_AUTO,
   DEMO_MODE,
   DEPLOYMENT_FLAVOR,
-  DUNE_API_KEY,
-  DUNE_POOLS_APY_URL,
   KNOWN_INITIATIVES_URL,
   LIQUITY_STATS_URL,
   SAFE_API_URL,
