+++ frontend/app/src/liquity-utils.ts
@@ -7,7 +7,7 @@ import type { Config as WagmiConfig } from "wagmi";
 import { DATA_REFRESH_INTERVAL, INTEREST_RATE_INCREMENT, INTEREST_RATE_MAX, INTEREST_RATE_MIN } from "@/src/constants";
 import { getCollateralContract, getContracts, getProtocolContract } from "@/src/contracts";
 import { dnum18, jsonStringifyWithDnum } from "@/src/dnum-utils";
-import { CHAIN_BLOCK_EXPLORER, DUNE_API_KEY, DUNE_POOLS_APY_URL } from "@/src/env";
+import { CHAIN_BLOCK_EXPLORER, LIQUITY_STATS_URL } from "@/src/env";
 import { getCollGainFromSnapshots, useContinuousBoldGains } from "@/src/liquity-stability-pool";
 import {
   useGovernanceStats,
@@ -81,90 +81,19 @@ export function getCollIndexFromSymbol(symbol: CollateralSymbol | null): CollInd
   return isCollIndex(collIndex) ? collIndex : null;
 }
 
-function vDunePoolsApyQuery(count: number) {
-  return v.pipe(
-    v.object({
-      result: v.object({
-        rows: v.pipe(
-          v.array(
-            v.object({
-              apr: v.number(),
-              avg_supply: v.number(),
-              bold_supply: v.number(),
-              collateral_type: v.string(),
-              day: v.string(),
-              period_apy: v.number(),
-              rewards: v.number(),
-              total_apr: v.number(),
-              total_rewards: v.number(),
-            }),
-          ),
-          v.length(count),
-        ),
-      }),
-    }),
-    v.transform(({ result }) => (
-      Object.fromEntries(
-        result.rows.map((row) => {
-          let symbol = row.collateral_type.toUpperCase();
-          if (symbol === "WETH") symbol = "ETH";
-          return isCollateralSymbol(symbol)
-            ? [symbol, {
-              apr: dn.from(row.apr, 18),
-              avgSupply: dn.from(row.avg_supply, 18),
-              boldSupply: dn.from(row.bold_supply, 18),
-              day: new Date(row.day),
-              rewards: dn.from(row.rewards, 18),
-              totalRewards: dn.from(row.total_rewards, 18),
-            }]
-            : null;
-        }).filter((v) => v !== null),
-      ) as Record<CollateralSymbol, {
-        apr: Dnum;
-        avgSupply: Dnum;
-        boldSupply: Dnum;
-        day: Date;
-        rewards: Dnum;
-        totalRewards: Dnum;
-      }>
-    )),
-  );
-}
-
 export function useEarnPool(collIndex: null | CollIndex) {
   const collateral = getCollToken(collIndex);
   const pool = useStabilityPool(collIndex ?? undefined);
+  const { data: stats } = useLiquityStats();
 
-  const aprs = useQuery({
-    queryKey: ["dunePoolsApy"],
-    queryFn: async () => {
-      if (!DUNE_POOLS_APY_URL || !DUNE_API_KEY) {
-        return null;
-      }
-      try {
-        const colls = getContracts().collaterals.length;
-        const response = await fetch(DUNE_POOLS_APY_URL, {
-          headers: { "X-Dune-API-Key": DUNE_API_KEY },
-        });
-        return v.parse(
-          vDunePoolsApyQuery(colls),
-          await response.json(),
-        );
-      } catch (e) {
-        console.error(e);
-        return null;
-      }
-    },
-    refetchInterval: 10 * 60 * 1000, // 10 minutes
-    enabled: Boolean(DUNE_POOLS_APY_URL && DUNE_API_KEY),
-  });
-
-  const apr = collateral?.symbol && aprs.data?.[collateral.symbol]?.apr;
+  const spAvgApy = (
+    collateral && stats?.branch[collateral.symbol]?.sp_apy_avg
+  ) ?? null;
 
   return {
     ...pool,
     data: {
-      apr: apr ?? null,
+      apr: spAvgApy === null ? null : dn.from(spAvgApy, 18),
       collateral,
       totalDeposited: pool.data?.totalDeposited ?? null,
     },
@@ -608,4 +537,51 @@ export async function getTroveOperationHints({
   });
 
   return { upperHint, lowerHint };
+}
+
+const StatsSchema = v.object({
+  total_bold_supply: v.string(),
+  total_debt_pending: v.string(),
+  total_coll_value: v.string(),
+  total_sp_deposits: v.string(),
+  total_value_locked: v.string(),
+  max_sp_apy: v.string(),
+  branch: v.record(
+    v.pipe(
+      v.string(),
+      v.transform((value) => {
+        value = value.toUpperCase();
+        if (value === "WETH") return "ETH";
+        return value;
+      }),
+    ),
+    v.object({
+      coll_active: v.string(),
+      coll_default: v.string(),
+      coll_price: v.string(),
+      sp_deposits: v.string(),
+      interest_accrual_1y: v.string(),
+      interest_pending: v.string(),
+      batch_management_fees_pending: v.string(),
+      debt_pending: v.string(),
+      coll_value: v.string(),
+      sp_apy: v.string(),
+      sp_apy_avg: v.optional(v.string()),
+      value_locked: v.string(),
+    }),
+  ),
+});
+
+export function useLiquityStats() {
+  return useQuery({
+    queryKey: ["liquity-stats"],
+    queryFn: async () => {
+      if (!LIQUITY_STATS_URL) {
+        throw new Error("LIQUITY_STATS_URL is not defined");
+      }
+      const response = await fetch(LIQUITY_STATS_URL);
+      return v.parse(StatsSchema, await response.json());
+    },
+    enabled: Boolean(LIQUITY_STATS_URL),
+  });
 }
